requirements.txt
apache-airflow-providers-http>=6.0.0
apache-airflow-providers-postgres>=5.0.0
apache-airflow-providers-smtp>=5.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.6
requests>=2.31.0
